+++ docs/changes.rst
@@ -9,6 +9,7 @@ Released on ? 2017.
 * Fixed quality checks on translation templates.
 * Added quality check to trigger on losing translation.
 * Add option to view pending suggestions from user.
+* Add option to automatically build component lists.
 
 weblate 2.12
 ------------